progress back end

+++ app/src/main/java/com/sidukov/weatherapp/domain/Weather.kt
@@ -6,7 +6,7 @@ data class Weather(
     val date: String,
     val image: Int,
     val temperature: Int,
-    val humidity: Float
+    val humidity: Int
 )
